# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Cinéphoria API
  version: 1.0.0
  description: Documentation de l’API REST Cinéphoria (Node.js, Express, PostgreSQL)
  contact:
    name: Redondo
    email: contact@cinephoria.fr
servers:
  - url: http://localhost:3000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /admin/dashboard:
    get:
      summary: Accède au tableau de bord de l’administrateur
      tags:
        - Administration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bienvenue sur le tableau de bord
        "401":
          description: Accès non autorisé
  /admin/dashboard/reservations:
    get:
      summary: Statistiques des réservations (MongoDB)
      tags:
        - Administration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Données de réservations sur 7 jours
  /admin/employees:
    post:
      summary: Crée un nouvel employé
      tags:
        - Administration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: employe@test.com
                role:
                  type: string
                  example: employe
      responses:
        "201":
          description: Employé créé
        "400":
          description: Erreur de validation
  "/admin/employees/{id}/reset-password":
    patch:
      summary: Réinitialise le mot de passe d’un employé
      tags:
        - Administration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID de l’employé
      responses:
        "200":
          description: Mot de passe réinitialisé
  /admin/films:
    post:
      summary: Ajoute un film
      tags:
        - Administration
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Film ajouté
  "/admin/films/{id}":
    patch:
      summary: Met à jour un film
      tags:
        - Administration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Film mis à jour
    delete:
      summary: Supprime un film
      tags:
        - Administration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Film supprimé
  /admin/seances:
    post:
      summary: Ajoute une séance
      tags:
        - Administration
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Séance ajoutée
  "/admin/seances/{id}":
    patch:
      summary: Met à jour une séance
      tags:
        - Administration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Séance mise à jour
    delete:
      summary: Supprime une séance
      tags:
        - Administration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Séance supprimée
  /admin/salles:
    post:
      summary: Ajoute une salle
      tags:
        - Administration
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Salle ajoutée
  "/admin/salles/{id}":
    patch:
      summary: Met à jour une salle
      tags:
        - Administration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Salle mise à jour
    delete:
      summary: Supprime une salle
      tags:
        - Administration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Salle supprimée
  /admin/cinemas:
    post:
      summary: Ajoute un cinéma
      tags:
        - Administration
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Cinéma ajouté
  "/admin/cinemas/{id}":
    patch:
      summary: Met à jour un cinéma
      tags:
        - Administration
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Cinéma mis à jour
    delete:
      summary: Supprime un cinéma
      tags:
        - Administration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cinéma supprimé
  /admin/users:
    get:
      summary: Liste tous les utilisateurs
      tags:
        - Administration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste renvoyée
  "/admin/users/{id}":
    patch:
      summary: Met à jour un utilisateur
      tags:
        - Administration
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Utilisateur mis à jour
    delete:
      summary: Supprime un utilisateur
      tags:
        - Administration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Utilisateur supprimé
...